updated budgetController logic to somewhat look right, at least with error messages and console logs. still needs updated query strings
